moved parsing cli args to config.js
